refactor(tabs): move /tab command ADTs to Tabs.Types; add _env_runTabCommand seam
