+++ src/PureClaw/Tabs/Types.hs
@@ -52,6 +52,10 @@ module PureClaw.Tabs.Types
   , conversationsOn
   , pruneDangling
   , relayModeFor
+    -- * @\/tab@ command ADTs (parsed form; shared by the dispatcher seam)
+  , TabKindArg (..)
+  , ForceMode (..)
+  , TabSlashCommand (..)
   ) where
 
 import Data.List.NonEmpty (NonEmpty)
@@ -317,3 +321,76 @@ pruneDangling tl cs =
 -- set, otherwise the supplied global default.
 relayModeFor :: ConversationKey -> RelayMode -> CursorState -> RelayMode
 relayModeFor k def cs = Map.findWithDefault def k (_cs_relay cs)
+
+-- ---------------------------------------------------------------------------
+-- @/tab@ command ADTs (parsed form)
+-- ---------------------------------------------------------------------------
+
+-- | A redacted enumeration of @TabKind@ as it appears in the @/tab new@
+-- command. Local to this module to avoid an import cycle through
+-- 'PureClaw.Handles.Tab' (which imports this module for 'SlashCommand'
+-- in its 'TabUnsupportedCommand' constructor). WU2 introduces this type
+-- alongside the @/tab@ command family; downstream WUs that need a
+-- 'PureClaw.Handles.Tab.TabKind' translate via a trivial total
+-- conversion at the handler layer (WU9).
+data TabKindArg
+  = TkaAi
+  | TkaProvider
+  | TkaHarness
+  | TkaShell
+  | TkaSsh
+  | TkaTmux
+  deriving stock (Show, Eq, Ord, Bounded, Enum)
+
+-- | Whether @/tab close N@ was passed the @--force@ flag.
+--
+-- For @KindAi@ tabs, @ForceYes@ skips the session archive on close
+-- (transcript deleted from disk). For non-AI tabs the close path is
+-- already destructive so the flag is a no-op semantically — the
+-- distinction is preserved here so 'executeSlashCommand' (WU9) can
+-- still echo what the user asked for.
+data ForceMode
+  = ForceNo
+  | ForceYes
+  deriving stock (Show, Eq, Ord, Bounded, Enum)
+
+-- | The @/tab@ command family (introduced by WU2 of Tabbed Chat #51).
+--
+-- Tab indices are stored as plain 'Int' rather than as
+-- 'PureClaw.Handles.Tab.TabIndex' to avoid an import cycle. The parser
+-- ('PureClaw.Routing.Parse.parseInput') validates the index against
+-- @_rc_maxTabs@ via @mkTabIndex@ before constructing these values, so
+-- callers may treat the contained 'Int' as well-formed for the
+-- configured cap; downstream handlers (WU9) are still expected to
+-- re-wrap via @mkTabIndex@ when crossing into 'TabIndex'-typed APIs.
+--
+-- @TabResumeCmd@ carries the validated 'SessionId' produced by
+-- @mkSessionId@ (WU2 smart constructor).
+--
+-- /tmux-style packing update:/ @TabNewCmd@ no longer carries an
+-- explicit target index — new tabs are always allocated at the lowest
+-- free slot. The constructor's payload is therefore just the optional
+-- kind keyword and the optional argument-text remainder.
+data TabSlashCommand
+  = TabNewCmd !(Maybe TabKindArg) !(Maybe Text)
+    -- ^ @\/tab new [\<kind\> [\<arg-text\>]]@. Index is allocated at
+    --   the lowest free slot by the handler (tmux-style packing). The
+    --   second field is the remainder of the line after the kind,
+    --   captured as a single 'Text' for the handler to split further.
+  | TabListCmd
+    -- ^ @\/tab list@ (and the @\/tabs@ alias).
+  | TabCloseCmd !Int !ForceMode
+    -- ^ @\/tab close \<N\> [--force]@. Remaining tabs are renumbered
+    --   down by one starting at @N+1@ so the registry is always packed
+    --   in the lowest slots (tmux @renumber-windows on@ model).
+  | TabFocusCmd !Int
+    -- ^ @\/tab focus \<N\>@ (functional alias of @\/N@).
+  | TabResumeCmd !SessionId
+    -- ^ @\/tab resume \<session-id\>@. Validation lives in
+    --   @mkSessionId@; rejection surfaces as
+    --   @ParseErrorInvalidSessionId@.
+  | TabRenameCmd !Int !Text
+    -- ^ @\/tab rename \<N\> \<name\>@. Parser captures the requested
+    --   name verbatim; @sanitizeTabName@ runs at handler time per
+    --   S10 so the user sees the rejection reason when applicable.
+  deriving stock (Show, Eq)
