+++ test/Agent/BackgroundSpec.hs
@@ -87,6 +87,7 @@ mkTestEnv p ch = do
     , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
     , _env_onTabsChanged = pure ()
     , _env_startHarness  = noStartHarness
+    , _env_runTabCommand = noRunTabCommand
     }
 
 -- | Like 'mkTestEnv' but with a REAL foreground session rooted under the
