+++ test/Agent/SlashCommandsSpec.hs
@@ -364,6 +364,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
 
     it "/new clears messages but keeps system prompt" $ do
@@ -530,6 +531,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdProvider ProviderList) ctx
@@ -584,6 +586,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdProvider ProviderList) ctx
@@ -634,6 +637,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdProvider (ProviderConfigure "badname")) ctx
@@ -684,6 +688,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdProvider (ProviderConfigure "ollama")) ctx
@@ -746,6 +751,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdProvider (ProviderConfigure "ollama")) ctx
@@ -799,6 +805,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTarget Nothing) ctx
@@ -848,6 +855,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTarget (Just "llama3")) ctx
@@ -908,6 +916,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTarget (Just "claude-code")) ctx
@@ -970,6 +979,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdMsg "claude-code-0" "list TODOs") ctx
@@ -1023,6 +1033,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdMsg "nonexistent" "hello") ctx
@@ -1076,6 +1087,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdMsg "cc-0" "test") ctx
@@ -1124,6 +1136,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
 
     it "/vault list with no vault → helpful message" $ do
@@ -1187,6 +1200,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1246,6 +1260,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1298,6 +1313,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
 
     it "/vault setup presents menu with passphrase option" $ withTempHome $ do
@@ -1341,6 +1357,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1395,6 +1412,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1455,6 +1473,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1517,6 +1536,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1568,6 +1588,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault VaultSetup) ctx
@@ -1685,6 +1706,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault (VaultDelete "todelete")) ctx
@@ -1739,6 +1761,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault (VaultDelete "keep")) ctx
@@ -1794,6 +1817,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdVault (VaultAdd "mykey")) ctx
@@ -1885,6 +1909,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env CmdHelp ctx
@@ -1944,6 +1969,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env CmdHelp ctx
@@ -1996,6 +2022,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = addMessage (textMessage User "hello") (emptyContext Nothing)
       ctx' <- executeSlashCommand env CmdHelp ctx
@@ -2043,6 +2070,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = addMessage (textMessage User "hello") (emptyContext Nothing)
       ctx' <- executeSlashCommand env (CmdBg "x") ctx
@@ -2171,6 +2199,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript (TranscriptRecent Nothing)) ctx
@@ -2223,6 +2252,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript TranscriptPath) ctx
@@ -2287,6 +2317,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript (TranscriptRecent Nothing)) ctx
@@ -2342,6 +2373,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript (TranscriptRecent Nothing)) ctx
@@ -2419,6 +2451,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript (TranscriptSearch "ollama")) ctx
@@ -2473,6 +2506,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript (TranscriptUnknown "badcmd")) ctx
@@ -2524,6 +2558,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript TranscriptPath) ctx
@@ -2573,6 +2608,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env CmdHelp ctx
@@ -2637,6 +2673,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
           ctx = emptyContext Nothing
       _ <- executeSlashCommand env (CmdTranscript (TranscriptRecent Nothing)) ctx
@@ -2718,6 +2755,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
 
     it "/agent list lists discovered agent names" $ withTempHome $ do
@@ -2786,6 +2824,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
 
     it "/agent info <name> shows files and frontmatter" $ withTempHome $ do
@@ -2952,6 +2991,7 @@ spec = do
             , _env_tabOutQ = error "8c.2 stub: _env_tabOutQ not exercised in this test"
             , _env_onTabsChanged = pure ()
             , _env_startHarness  = noStartHarness
+            , _env_runTabCommand = noRunTabCommand
             }
 
     it "/session new writes session.json on disk and returns a confirmation" $ withTempHome $ do
