+++ test/Support/AgentEnv.hs
@@ -67,4 +67,5 @@ mkTestAgentEnv = do
     , _env_tabOutQ           = _ts_tabOutQ ts
     , _env_onTabsChanged     = _ts_onTabsChanged ts
     , _env_startHarness      = noStartHarness
+    , _env_runTabCommand     = noRunTabCommand
     }
