+++ test/Tabs/WiringSpec.hs
@@ -326,6 +326,7 @@ mkTabbedEnvChan sessionsDir (channel, clog) = do
         , _env_tabOutQ           = _ts_tabOutQ ts
         , _env_onTabsChanged     = _ts_onTabsChanged ts
         , _env_startHarness      = noStartHarness
+        , _env_runTabCommand     = noRunTabCommand
         }
   pure (TabbedHarness env clog runnersTracker)
